Code refactor, incorporating geospatial
